remove pygame and ssl test env
move env init files to remove unneeded dependencies
