+++ envs/rc_gym/Utils/__init__.py
@@ -1,3 +1,2 @@
 from rc_gym.Utils.Utils import *
 from rc_gym.Utils.Render import *
-from rc_gym.Utils.Render_gym import *
